% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{complementarySketching}
\alias{complementarySketching}
\title{Main function implementing the complementary sketching algorithm in Gao and Wang (2020)}
\usage{
complementarySketching(X1, X2, y1, y2, sparse = TRUE, sigma = NULL)
}
\arguments{
\item{X1}{design matrix of the first linear regression model}

\item{X2}{design matrix of the second linear regression model}

\item{y1}{response vector of the first linear regression model}

\item{y2}{response vector of the second linear regression model}

\item{sparse}{whether to test a sparse difference in the regression coefficients of the two models, default is TRUE}

\item{sigma}{noise standard deviation of the regression models; if unknown, set to NULL and will be estimated.}
}
\value{
a list containing the test statistics and test results (0 for nonrejection and 1 for rejection).
}
\description{
From the model y1 = X1 beta1 + eps1 and y2 = X2 beta2 + eps2, where X1 and X2 are n1 x p and n2 x p design matrices respectively, we test the null H0: beta1 = beta2 against the alternative that beta1 - beta2 is nonzero (if sparse=FALSE) or beta1 - beta2 is nonzero and has at most sqrt(p) (if sparse=TRUE). The test is performed using the complementary sketching algorithm from Gao and Wang (2020).
}
\examples{
# problem parameters
n1 <- 600 # sample size of first sample
n2 <- 600 # sample size of second sample
p <- 800 # dimension of covariates
k <- 10 # sparsity of difference of the two regression coefficients
rho <- 1 # difference in l_2 norm of the two regression coefficients
# generate design matrices
X1 <- matrix(rnorm(n1 * p), n1, p)
X2 <- matrix(rnorm(n2 * p), n2, p)
# generate regression coefficients
beta1 <- rnorm(p)
theta <- c(rnorm(k), rep(0, p-k)) 
theta <- theta / sqrt(sum(theta^2)) * rho
beta2 <- beta1 + theta
# generate response vectors
y1 <- X1 \%*\% beta1 + rnorm(n1)
y2 <- X2 \%*\% beta2 + rnorm(n2)
# test for difference in beta1 and beta2
complementarySketching(X1,X2,y1,y2)
}
\references{
Gao, F. and Wang, T. (2020) Two-sample testing of high-dimensional linear regression coefficients via complementary sketching. Preprint, arxiv:2011.13624.
}
